Move oauth providers to separate files
